requirements.txt
aiohttp
aiohttp_jinja2
pyhamcrest
pytest
pytest-aiohttp
lxml
html5lib